requirements.txt
botocore==1.23.38
jmespath==0.10.0
kafka-python==2.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
urllib3==1.26.7